# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: Alexa app
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /anything
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /login:
    # binds a127 app logic to a route
    x-swagger-router-controller: alexa
    post:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: login
      parameters:
        - name: data
          in: body
          description: The name of the person to whom to say hello
          schema:
            # a pointer to a definition
            $ref: "#/definitions/loginrequest"
          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/loginresponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /weather:
    # binds a127 app logic to a route
    x-swagger-router-controller: alexa
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: weather
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/weatherresponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  loginresponse:
    properties:
      message:
        type: string
      status:
        type: string
      outputDate:
        type: string
  loginrequest:
    properties:
      inputDate:
        type: string
      id:
        type: string
      pwd:
        type: string
  weatherresponse:
    properties:
      message:
        type: string
          
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string